Finish CRUD of flight entity and create 2 custom exceptions

+++ src/main/java/com/example/flight/services/FlightService.java
@@ -1,13 +1,13 @@
 package com.example.flight.services;
 
 import com.example.flight.entities.Flight;
+import com.example.flight.exceptions.FlightNotFoundException;
 import com.example.flight.repositories.FlightRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-//Todo fix and create the correct exceptions for getFlightById and UpdateFlight
 @Service
 public class FlightService {
 
@@ -28,20 +28,21 @@ public class FlightService {
 
     public Flight getFlightById(Long id) {
         return repository.findById(id)
-                .orElseThrow(RuntimeException::new);
+                .orElseThrow(() -> new FlightNotFoundException(id));
     }
 
-//    public Flight updateFlight(Flight flight, Long id) {
-//        return repository.findById(id)
-//                .map(flightFound -> {
-//                    flightFound.setAirport(flight.getAirport());
-//                    flightFound.setAvailable(flight.isAvailable());
-//                    return repository.save(flightFound);
-//                })
-//                .orElseThrow(RuntimeException::new);
-//    }
+    public Flight updateFlight(Flight flight, Long id) {
+        return repository.findById(id)
+                .map(flightFound -> {
+                    flightFound.setAvailable(flight.isAvailable());
+                    return repository.save(flightFound);
+                })
+                .orElseThrow(() -> new FlightNotFoundException(id));
+    }
 
     public void deleteFlight(Long id) {
-        repository.deleteById(id);
+        if (getFlightById(id) != null) {
+            repository.deleteById(id);
+        }
     }
 }
